requirements.txt
python-decouple==3.8
python-telegram-bot==21.10